api-macros: Split request code generation into more files
